requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.20
django-import-export==2.7.1
django-simpleui==2023.8.28
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==4.8.3
inflection==0.5.1
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
pyparsing==3.1.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
zope.event==4.6
zope.interface==5.5.2
